split interfaces into http and websocket headers
