+++ include/cpprealm/app.hpp
@@ -27,8 +27,9 @@
 #include <cpprealm/internal/bridge/sync_manager.hpp>
 #include <cpprealm/internal/bridge/sync_session.hpp>
 #include <cpprealm/internal/bridge/utils.hpp>
+#include <cpprealm/networking/http.hpp>
 #include <cpprealm/networking/networking.hpp>
-#include <cpprealm/networking/platform_networking.hpp>
+#include <cpprealm/networking/websocket.hpp>
 
 #include <future>
 #include <utility>
