+++ include/cpprealm/internal/networking/utils.hpp
@@ -19,6 +19,8 @@
 #ifndef CPPREALM_NETWORKING_UTILS_HPP
 #define CPPREALM_NETWORKING_UTILS_HPP
 
+#include <cpprealm/networking/websocket.hpp>
+
 #include <optional>
 
 namespace realm {
@@ -31,23 +33,21 @@ namespace realm {
     }
 
     namespace networking {
-        struct default_transport_configuration;
         struct request;
         struct response;
-        struct websocket_endpoint;
     }
 }
 
 namespace realm::internal::networking {
     ::realm::networking::request to_request(const ::realm::app::Request&);
-    ::realm::app::Request to_core_request(const ::realm::networking::request&);
+//    ::realm::app::Request to_core_request(const ::realm::networking::request&);
 
-    ::realm::networking::response to_response(const ::realm::app::Response&);
+//    ::realm::networking::response to_response(const ::realm::app::Response&);
     ::realm::app::Response to_core_response(const ::realm::networking::response&);
 
-    ::realm::sync::WebSocketEndpoint to_core_websocket_endpoint(const ::realm::networking::websocket_endpoint& ep,
-                                                                const std::optional<::realm::networking::default_transport_configuration>& config);
-    ::realm::networking::websocket_endpoint to_websocket_endpoint(const ::realm::sync::WebSocketEndpoint& ep);
+    ::realm::sync::WebSocketEndpoint to_core_websocket_endpoint(const ::realm::networking::sync_socket_provider::websocket_endpoint& ep,
+                                                                const std::optional<::realm::networking::default_socket_provider::configuration>& config);
+    ::realm::networking::sync_socket_provider::websocket_endpoint to_websocket_endpoint(const ::realm::sync::WebSocketEndpoint& ep);
 } //namespace realm::internal::networking
 
 #endif //CPPREALM_NETWORKING_UTILS_HPP
