+++ src/CMakeLists.txt
@@ -54,7 +54,8 @@ set(SOURCES
     cpprealm/internal/networking/shims.cpp
     cpprealm/internal/networking/utils.cpp
     cpprealm/internal/scheduler/realm_core_scheduler.cpp
-    cpprealm/networking/platform_networking.cpp
+    cpprealm/networking/http.cpp
+    cpprealm/networking/websocket.cpp
     cpprealm/schedulers/default_scheduler.cpp
     cpprealm/logger.cpp
     cpprealm/sdk.cpp) # REALM_SOURCES
@@ -118,7 +119,8 @@ set(HEADERS
     ../include/cpprealm/internal/scheduler/realm_core_scheduler.hpp
     ../include/cpprealm/schedulers/default_scheduler.hpp
     ../include/cpprealm/networking/networking.hpp
-    ../include/cpprealm/networking/platform_networking.hpp
+    ../include/cpprealm/networking/http.hpp
+    ../include/cpprealm/networking/websocket.hpp
     ../include/cpprealm/logger.hpp
     ../include/cpprealm/notifications.hpp
     ../include/cpprealm/rbool.hpp
