+++ src/cpprealm/internal/network/network_transport.cpp
@@ -20,7 +20,7 @@
 #include <realm/util/config.h>
 #endif
 #include <cpprealm/app.hpp>
-#include <cpprealm/networking/platform_networking.hpp>
+#include <cpprealm/networking/http.hpp>
 #include <cpprealm/networking/networking.hpp>
 
 #include <realm/object-store/sync/generic_network_transport.hpp>
