+++ src/cpprealm/internal/networking/utils.cpp
@@ -1,5 +1,5 @@
 #include <cpprealm/internal/networking/utils.hpp>
-#include <cpprealm/networking/platform_networking.hpp>
+#include <cpprealm/networking/http.hpp>
 
 #include <realm/object-store/sync/generic_network_transport.hpp>
 #include <realm/sync/protocol.hpp>
@@ -8,6 +8,7 @@
 #include <realm/util/uri.hpp>
 
 namespace realm::internal::networking {
+
     ::realm::networking::request to_request(const ::realm::app::Request& core_request) {
         ::realm::networking::request req;
         req.method = static_cast<::realm::networking::http_method>(core_request.method);
@@ -17,25 +18,7 @@ namespace realm::internal::networking {
         req.body = core_request.body;
         return req;
     }
-    ::realm::app::Request to_core_request(const ::realm::networking::request& req) {
-        ::realm::app::Request core_request;
-        core_request.method = static_cast<::realm::app::HttpMethod>(req.method);
-        core_request.url = req.url;
-        core_request.timeout_ms = req.timeout_ms;
-        core_request.headers = req.headers;
-        core_request.body = req.body;
-        return core_request;
-    }
 
-    ::realm::networking::response to_response(const ::realm::app::Response& core_response) {
-        ::realm::networking::response req;
-        req.http_status_code = core_response.http_status_code;
-        req.custom_status_code = core_response.custom_status_code;
-        req.headers = core_response.headers;
-        req.client_error_code = core_response.client_error_code;
-        req.body = core_response.body;
-        return req;
-    }
     ::realm::app::Response to_core_response(const ::realm::networking::response& req) {
         ::realm::app::Response core_response;
         core_response.http_status_code = req.http_status_code;
@@ -62,8 +45,8 @@ namespace realm::internal::networking {
         REALM_TERMINATE("Unrecognized websocket protocol");
     }
 
-    ::realm::sync::WebSocketEndpoint to_core_websocket_endpoint(const ::realm::networking::websocket_endpoint& ep,
-                                                                const std::optional<::realm::networking::default_transport_configuration>& config) {
+    ::realm::sync::WebSocketEndpoint to_core_websocket_endpoint(const ::realm::networking::sync_socket_provider::websocket_endpoint& ep,
+                                                                const std::optional<::realm::networking::default_socket_provider::configuration>& config) {
         ::realm::sync::WebSocketEndpoint core_ep;
         auto uri = util::Uri(ep.url);
         auto protocol = to_protocol_envelope(uri.get_scheme());
@@ -107,8 +90,8 @@ namespace realm::internal::networking {
         return core_ep;
     }
 
-    ::realm::networking::websocket_endpoint to_websocket_endpoint(const ::realm::sync::WebSocketEndpoint& core_ep) {
-        ::realm::networking::websocket_endpoint ep;
+    ::realm::networking::sync_socket_provider::websocket_endpoint to_websocket_endpoint(const ::realm::sync::WebSocketEndpoint& core_ep) {
+        ::realm::networking::sync_socket_provider::websocket_endpoint ep;
         ep.protocols = core_ep.protocols;
         const auto& port = core_ep.proxy ? core_ep.proxy->port : core_ep.port;
         ep.url = util::format("%1://%2:%3%4", core_ep.is_ssl ? "wss" : "ws", core_ep.address, port, core_ep.path);
