+++ tests/admin_utils.cpp
@@ -20,7 +20,7 @@
 #include <sstream>
 #include <thread>
 #include <iostream>
-#include <cpprealm/networking/platform_networking.hpp>
+#include <cpprealm/networking/http.hpp>
 #include <cpprealm/internal/networking/utils.hpp>
 
 #include "admin_utils.hpp"
